♻️refactor: 3개 식물 셀 -> 단일 식물 셀 리팩토링
