--- LeafLog/LeafLog/View/HomeView/SinglePlantShelfCell.swift
@@ -1,75 +1,0 @@
-//
-//  SinglePlantShelfCell.swift
-//  LeafLog
-//
-//  Created by t2025-m0143 on 4/15/26.
-//
-
-import UIKit
-import SnapKit
-import Then
-
-class SinglePlantShelfCell: UICollectionViewCell {
-    private let plant = UIImageView()
-    
-    private let card = PlantLabelCardView()
-    
-    private let shelf = SeparateBar().then {
-        $0.snp.makeConstraints {
-            $0.height.equalTo(6)
-        }
-    }
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        setLayout()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
-
-extension SinglePlantShelfCell {
-    
-    private func setLayout() {
-        contentView.addSubview(plant)
-        contentView.addSubview(shelf)
-        contentView.addSubview(card)
-        
-        plant.snp.makeConstraints {
-            $0.top.equalToSuperview()
-            $0.horizontalEdges.equalToSuperview().inset(16)
-            $0.height.equalTo(plant.snp.width)
-        }
-        
-        shelf.snp.makeConstraints {
-            $0.top.equalTo(plant.snp.bottom)
-            $0.horizontalEdges.equalToSuperview()
-        }
-        
-        card.snp.makeConstraints {
-            $0.top.equalTo(shelf.snp.bottom).offset(12)
-            $0.horizontalEdges.equalToSuperview().inset(8)
-        }
-    }
-}
-
-extension SinglePlantShelfCell {
-    func configure(_ data: HomeView.SingleShelfPlant) {
-        if let target = data.isAddButton {
-            card.isHidden = true
-            if target {
-                plant.image = .plantAdd
-            } else {
-                plant.image = nil
-            }
-        } else {
-            plant.image = UIImage(named: data.category?.defaultImageAssetName ?? "")
-            card.nameLabel.text = data.name ?? ""
-            card.recentDayLabel.text = "\(String(describing: data.daysFromLastWatering ?? 0))일 전"
-            card.nextDayLabel.text = "\(String(describing: data.daysToNextWatering ?? 0))일"
-            card.waterButton.isSelected = data.didWater ?? false
-        }
-    }
-}
